refactor: Update import paths for adblocker module
